feat: format Nix files and respect DRY_RUN and VERBOSE variables.

+++ module/yabai/default.nix
@@ -1,27 +1,35 @@
-{ config, pkgs, lib, ... }:
-
 {
-	home.packages = [
-		pkgs.yabai
-	];
-	
-	home.file = {
-		"yabairc" = {
-			source = config.lib.file.mkOutOfStoreSymlink ./yabairc;
-			target = ".config/yabai/yabairc";
-		};
-	};
+  config,
+  pkgs,
+  lib,
+  ...
+}: {
+  home.packages = [
+    pkgs.yabai
+  ];
 
-	home.activation = {
-		setupYabai = lib.hm.dag.entryAfter ["writeBoundary"] 
-			"(
+  home.file = {
+    "yabairc" = {
+      source = config.lib.file.mkOutOfStoreSymlink ./yabairc;
+      target = ".config/yabai/yabairc";
+    };
+  };
+
+  home.activation = {
+    setupYabai =
+      lib.hm.dag.entryAfter ["writeBoundary"]
+      "(
 				PATH=$(realpath $HOME)/.nix-profile/bin:/usr/bin:/bin:$PATH
 				echo -n 'Do you want to setup the yabai scripting addition (superuser privileges required)? [y/n]: '
 				read -r reply
 				[ ! -t 0 ] && echo
 				if [[ $reply =~ ^[Yy]$ ]]; then
-					echo \"$(whoami) ALL=(root) NOPASSWD: sha256:$(shasum -a 256 $(which yabai) | cut -d ' ' -f 1) $(which yabai) --load-sa\" | sudo tee /private/etc/sudoers.d/yabai > /dev/null
+					if [[ -v DRY_RUN ]]; then
+						echo 'echo \"$(whoami) ALL=(root) NOPASSWD: sha256:$(shasum -a 256 $(which yabai) | cut -d ' ' -f 1) $(which yabai) --load-sa\" | sudo tee /private/etc/sudoers.d/yabai > /dev/null'
+					else
+						echo \"$(whoami) ALL=(root) NOPASSWD: sha256:$(shasum -a 256 $(which yabai) | cut -d ' ' -f 1) $(which yabai) --load-sa\" | sudo tee /private/etc/sudoers.d/yabai > /dev/null
+					fi
 				fi
 			 )";
-	};
+  };
 }
